requirements.txt
pyotp~=2.4.0
requests~=2.22.0
tweepy~=3.9.0
beautifulsoup4~=4.8.2
fuzzywuzzy~=0.18.0
Pillow~=7.0.0
mysql~=0.0.2
tensorflow~=2.3.0
Keras~=2.4.3
numpy~=1.18.1
pandas~=1.0.1
scikit-learn~=0.22.1
matplotlib~=3.1.3
scipy~=1.4.1
discord.py~=1.4.1
steam_community_market~=1.2.3
mysql-connector~=2.2.9
lxml~=4.5.2